requirements.txt
mediapipe~=0.10.9
python-statemachine[diagrams]
pydot
#sometimes should be installed with sudo apt-get install graphviz or via https://graphviz.org/download/
Graphviz
requests-cache
retry-requests
# only on no PI OS :
# gpiozero
